pass 1
Idea was to save PVs count in local storage and then send them when online
But the problem is we can communicate outside of cached stuff only when we are refreshing. In cached state it can't load GA's JS.

Next step would be to try caching GA's JS in manifest to see after that we can send tracking data.
It comes with the problem that GA's JS gets cached till the time we update our manifest, which we can counterattack by making cache manifest invalidate every week.

This way users can continue to use cached site when they are offline and they will get update every week at max.